+++ tests/testthat/test-get_device_info.R
@@ -14,6 +14,22 @@ with_mock_dir("apsystems", {
   })
 })
 
+with_mock_dir("fronius", {
+  test_that("get_device_info() works with a single device from Fronius", {
+    skip_on_cran()
+    expect_error(
+      get_device_info(device_ip = "fronius.local", model = "Fronius"),
+      NA)
+    fronius_info <-  get_device_info(device_ip = "fronius.local", model = "Fronius")
+    expect_true(is.data.frame(fronius_info))
+    expect_equal(
+      names(fronius_info),
+      c("device_id", "last_report", "CustomName", "DT", "ErrorCode", "PVPower", "Show", "StatusCode", "UniqueID")
+      )
+    expect_equal(nrow(fronius_info), 1L)
+  })
+})
+
 with_mock_dir("apsystems", {
   test_that("get_device_info() works with multiple devices from APSystems", {
     skip_on_cran()