+++ NEWS.md
@@ -1,10 +1,11 @@
 # microinverterdata 0.1.2
 
-* Add support for 'Fronius' inverters
+* Add mocked http data for APSystems and Fronius
+* Add support for 'Fronius' inverters (`get_output_data()` only)
 
 # microinverterdata 0.1.1
 
-* Add support to Enphase Envoy-S inverters
+* Add support to Enphase Envoy-S inverters (`get_output_data()` only)
 * manage device unreachable error
 
 # microinverterdata 0.1.0